refactor keychains
