+++ core/src/wallet/bitcoin/keychains/P2PKHBitcoinLikeKeychain.cpp
@@ -29,18 +29,6 @@
  *
  */
 #include "P2PKHBitcoinLikeKeychain.hpp"
-#include <cereal/cereal.hpp>
-#include <cereal/archives/binary.hpp>
-#include <cereal/types/set.hpp>
-#include <boost/iostreams/device/array.hpp>
-#include <boost/iostreams/stream.hpp>
-#include <fmt/format.h>
-#include <src/api/DynamicObject.hpp>
-#include "../../../api/BitcoinLikeAddress.hpp"
-#include "../../../api/Configuration.hpp"
-#include "../../../utils/DerivationPath.hpp"
-#include "../../../collections/strings.hpp"
-#include <debug/Benchmarker.h>
 #include <bitcoin/BitcoinLikeAddress.hpp>
 
 namespace ledger {
@@ -51,138 +39,9 @@ namespace ledger {
                                                            int account,
                                                            const std::shared_ptr<api::BitcoinLikeExtendedPublicKey> &xpub,
                                                            const std::shared_ptr<Preferences> &preferences)
-                : BitcoinLikeKeychain(configuration, params, account, preferences) {
-            _xpub = xpub;
-
-            {
-                auto localPath = getDerivationScheme().getSchemeTo(DerivationSchemeLevel::NODE)
-                        .setAccountIndex(getAccountIndex())
-                        .setCoinType(getCurrency().bip44CoinType)
-                        .setNode(RECEIVE).getPath();
-                _publicNodeXpub = std::static_pointer_cast<BitcoinLikeExtendedPublicKey>(_xpub)->derive(localPath);
-            }
-            {
-                auto localPath = getDerivationScheme().getSchemeTo(DerivationSchemeLevel::NODE)
-                                                      .setAccountIndex(getAccountIndex())
-                                                      .setCoinType(getCurrency().bip44CoinType)
-                                                      .setNode(CHANGE).getPath();
-                _internalNodeXpub = std::static_pointer_cast<BitcoinLikeExtendedPublicKey>(_xpub)->derive(localPath);
-            }
-
-            // Try to restore the state from preferences
-            auto state = preferences->getData("state", {});
-            if (!state.empty()) {
-                boost::iostreams::array_source my_vec_source(reinterpret_cast<char*>(&state[0]), state.size());
-                boost::iostreams::stream<boost::iostreams::array_source> is(my_vec_source);
-                ::cereal::BinaryInputArchive archive(is);
-                archive(_state);
-            } else {
-                _state.maxConsecutiveReceiveIndex = 0;
-                _state.maxConsecutiveChangeIndex = 0;
-                _state.empty = true;
-            }
-            _observableRange = (uint32_t) configuration->getInt(api::Configuration::KEYCHAIN_OBSERVABLE_RANGE).value_or(20);
+                : CommonBitcoinLikeKeychains(configuration, params, account, xpub, preferences)
+        {
             getAllObservableAddresses(0, _observableRange);
-        }
-
-        bool P2PKHBitcoinLikeKeychain::markPathAsUsed(const DerivationPath &p) {
-            DerivationPath path(p);
-            if (path.getParent().getLastChildNum() == 0) {
-                if (path.getLastChildNum() < _state.maxConsecutiveReceiveIndex ||
-                    _state.nonConsecutiveReceiveIndexes.find(path.getLastChildNum()) != _state.nonConsecutiveReceiveIndexes.end()) {
-                    return false;
-                } else {
-                    if (path.getLastChildNum() == _state.maxConsecutiveReceiveIndex)
-                        _state.maxConsecutiveReceiveIndex += 1;
-                    else
-                        _state.nonConsecutiveReceiveIndexes.insert(path.getLastChildNum());
-                    _state.empty = false;
-                    saveState();
-                    getAllObservableAddresses(path.getLastChildNum(), path.getLastChildNum() + _observableRange);
-                    return true;
-                }
-            } else {
-                if (path.getLastChildNum() < _state.maxConsecutiveChangeIndex ||
-                    _state.nonConsecutiveChangeIndexes.find(path.getLastChildNum()) != _state.nonConsecutiveChangeIndexes.end()) {
-                    return false;
-                } else {
-                    if (path.getLastChildNum() == _state.maxConsecutiveChangeIndex)
-                        _state.maxConsecutiveChangeIndex += 1;
-                    else
-                        _state.nonConsecutiveChangeIndexes.insert(path.getLastChildNum());
-                    _state.empty = false;
-                    saveState();
-                    getAllObservableAddresses(path.getLastChildNum(), path.getLastChildNum() + _observableRange);
-                    return true;
-                }
-            }
-        }
-
-        BitcoinLikeKeychain::Address P2PKHBitcoinLikeKeychain::getFreshAddress(BitcoinLikeKeychain::KeyPurpose purpose) {
-            return derive(purpose, (purpose == KeyPurpose::RECEIVE ? _state.maxConsecutiveReceiveIndex : _state.maxConsecutiveChangeIndex));
-        }
-
-        bool P2PKHBitcoinLikeKeychain::isEmpty() const {
-            return _state.empty;
-        }
-
-        std::vector<BitcoinLikeKeychain::Address> P2PKHBitcoinLikeKeychain::getAllObservableAddresses(uint32_t from, uint32_t to) {
-            auto length = to - from;
-            std::vector<BitcoinLikeKeychain::Address> result;
-            result.reserve((length + 1) * 2);
-            for (auto i = 0; i <= length; i++) {
-                result.push_back(derive(KeyPurpose::RECEIVE, from + i));
-                result.push_back(derive(KeyPurpose::CHANGE, from + i));
-            }
-            return result;
-        }
-
-        std::vector<BitcoinLikeKeychain::Address>
-        P2PKHBitcoinLikeKeychain::getFreshAddresses(BitcoinLikeKeychain::KeyPurpose purpose, size_t n) {
-            auto startOffset = (purpose == KeyPurpose::RECEIVE) ? _state.maxConsecutiveReceiveIndex : _state.maxConsecutiveChangeIndex;
-            std::vector<BitcoinLikeKeychain::Address> result(n);
-            for (auto i = 0; i < n; i++) {
-                result[i] = derive(purpose, startOffset + i);
-            }
-            return result;
-        }
-
-        Option<BitcoinLikeKeychain::KeyPurpose>
-        P2PKHBitcoinLikeKeychain::getAddressPurpose(const std::string &address) const {
-            return getAddressDerivationPath(address).flatMap<KeyPurpose>([] (const std::string& p) {
-                DerivationPath derivation(p);
-                if (derivation.getDepth() > 1) {
-                    return Option<KeyPurpose>(derivation.getParent().getLastChildNum() == 0 ? KeyPurpose::RECEIVE : KeyPurpose::CHANGE);
-                } else {
-                    return Option<KeyPurpose>();
-                }
-            });
-        }
-
-        Option<std::string> P2PKHBitcoinLikeKeychain::getAddressDerivationPath(const std::string &address) const {
-            auto path = getPreferences()->getString(fmt::format("address:{}", address), "");
-            if (path.empty()) {
-                return Option<std::string>();
-            } else {
-                auto derivation = DerivationPath(getExtendedPublicKey()->getRootPath()) + DerivationPath(path);
-                return Option<std::string>(derivation.toString());
-            }
-        }
-
-        std::vector<BitcoinLikeKeychain::Address>
-        P2PKHBitcoinLikeKeychain::getAllObservableAddresses(BitcoinLikeKeychain::KeyPurpose purpose, uint32_t from,
-                                                            uint32_t to) {
-            auto maxObservableIndex = (purpose == KeyPurpose::CHANGE ? _state.maxConsecutiveChangeIndex + _state.nonConsecutiveChangeIndexes.size() : _state.maxConsecutiveReceiveIndex + _state.nonConsecutiveReceiveIndexes.size()) + _observableRange;
-            auto length = std::min<size_t >(to - from, maxObservableIndex - from);
-            std::vector<BitcoinLikeKeychain::Address> result(length +1);
-            for (auto i = 0; i <= length; i++) {
-                if (purpose == KeyPurpose::RECEIVE) {
-                    result.push_back(derive(KeyPurpose::RECEIVE, from + i));
-                } else {
-                    result.push_back(derive(KeyPurpose::CHANGE, from + i));
-                }
-            }
-            return result;
         }
 
         BitcoinLikeKeychain::Address P2PKHBitcoinLikeKeychain::derive(KeyPurpose purpose, off_t index) {
@@ -213,54 +72,11 @@ namespace ledger {
             return std::dynamic_pointer_cast<BitcoinLikeAddress>(BitcoinLikeAddress::parse(address, getCurrency(), Option<std::string>(localPath)));
         }
 
-        void P2PKHBitcoinLikeKeychain::saveState() {
-            while (_state.nonConsecutiveReceiveIndexes.find(_state.maxConsecutiveReceiveIndex) != _state.nonConsecutiveReceiveIndexes.end()) {
-                _state.nonConsecutiveReceiveIndexes.erase(_state.maxConsecutiveReceiveIndex);
-                _state.maxConsecutiveReceiveIndex += 1;
-            }
-            while (_state.nonConsecutiveChangeIndexes.find(_state.maxConsecutiveChangeIndex) != _state.nonConsecutiveChangeIndexes.end()) {
-                _state.nonConsecutiveChangeIndexes.erase(_state.maxConsecutiveChangeIndex);
-                _state.maxConsecutiveChangeIndex += 1;
-            }
-            std::stringstream is;
-            ::cereal::BinaryOutputArchive archive(is);
-            archive(_state);
-            auto savedState = is.str();
-            getPreferences()->edit()->putData("state", std::vector<uint8_t>((const uint8_t *)savedState.data(),(const uint8_t *)savedState.data() + savedState.size()))->commit();
-        }
-
-        std::shared_ptr<api::BitcoinLikeExtendedPublicKey> P2PKHBitcoinLikeKeychain::getExtendedPublicKey() const {
-            return _xpub;
-        }
-
-        std::string P2PKHBitcoinLikeKeychain::getRestoreKey() const {
-            return _xpub->toBase58();
-        }
-
-        int32_t P2PKHBitcoinLikeKeychain::getObservableRangeSize() const {
-            return _observableRange;
-        }
-
-        bool P2PKHBitcoinLikeKeychain::contains(const std::string &address) const {
-            return getAddressDerivationPath(address).nonEmpty();
-        }
-
         Option<std::string>
         P2PKHBitcoinLikeKeychain::getHash160DerivationPath(const std::vector<uint8_t> &hash160) const {
             const auto& params = getCurrency().bitcoinLikeNetworkParameters.value();
             BitcoinLikeAddress address(getCurrency(), hash160, params.P2PKHVersion);
             return getAddressDerivationPath(address.toBase58());
         }
-
-        Option<std::vector<uint8_t>> P2PKHBitcoinLikeKeychain::getPublicKey(const std::string &address) const {
-            auto path = getPreferences()->getString(fmt::format("address:{}", address), "");
-            if (path.empty()) {
-                Option<std::vector<uint8_t>>();
-            }
-            return Option<std::vector<uint8_t>>(_xpub->derivePublicKey(path));
-        }
-
     }
 }
-
-CEREAL_CLASS_VERSION(ledger::core::P2PKHKeychainPersistentState, 0);
